Refactored code, deleted comments, renamed packages.
