+++ med-web-app-front/src/components/messenger_components/recipient.msg.component.js
