+++ med-web-app-front/src/components/messenger_components/sender.msg.component.js
