+++ med-web-app-front/src/components/user-card.component.js
@@ -1,7 +1,6 @@
 import React, {Component} from "react";
 import '../styles/Search.css'
 import {TableCell, withStyles} from "@material-ui/core";
-// import {Link} from '@material-ui/core';
 import { Link } from "react-router-dom";
 const useStyles = theme => ({
     root: {
