+++ med-web-app-front/src/services/auth-header.js
@@ -2,7 +2,7 @@ export default function authHeader() {
     const user = JSON.parse(localStorage.getItem('user'));
 
     if (user && user.token) {
-        return { Authorization: 'Bearer ' + user.token };
+        return {Authorization: 'Bearer ' + user.token};
     } else {
         return {};
     }
