+++ med-web-app-front/src/services/pipeline.service.js
@@ -6,15 +6,15 @@ const API_URL = process.env.REACT_APP_API_URL + '/api/pipelines/';
 class PipelineService {
 
     getAllPipelines() {
-        return axios.get(API_URL + 'all', { headers: authHeader() });
+        return axios.get(API_URL + 'all', {headers: authHeader()});
     }
 
     savePipeline(jsonConfiguration, description) {
-        return axios.post(API_URL + 'save', {jsonConfiguration, description},{ headers: authHeader() });
+        return axios.post(API_URL + 'save', {jsonConfiguration, description}, {headers: authHeader()});
     }
 
     deletePipeline(pipelineId) {
-        return axios.delete(API_URL + pipelineId,{ headers: authHeader() });
+        return axios.delete(API_URL + pipelineId, {headers: authHeader()});
     }
 }
