+++ med-web-app-front/src/services/pipelinejob.service.js
@@ -6,15 +6,15 @@ const API_URL = process.env.REACT_APP_API_URL + '/api/pipelinejobs/';
 class PipelineJobService {
 
     getPipelineJobsForUser(username) {
-        return axios.get(API_URL + username, { headers: authHeader() });
+        return axios.get(API_URL + username, {headers: authHeader()});
     }
 
     sendRequestForPipelineJob(username, pipelineId, fileId) {
-        return axios.post(API_URL + username, {pipelineId, fileId},{ headers: authHeader() });
+        return axios.post(API_URL + username, {pipelineId, fileId}, {headers: authHeader()});
     }
 
     deletePipelineJob(username, pipelineJobId, fileId) {
-        return axios.delete(API_URL + username + "/" + pipelineJobId + "/" + fileId,{ headers: authHeader() });
+        return axios.delete(API_URL + username + "/" + pipelineJobId + "/" + fileId, {headers: authHeader()});
     }
 }
