+++ med-web-app-front/src/services/test.service.js
@@ -4,22 +4,21 @@ import authHeader from './auth-header';
 const API_URL = process.env.REACT_APP_API_URL + '/api/test/';
 
 
-
 class TestService {
     getPublicContent() {
         return axios.get(API_URL + 'all');
     }
 
     getUserBoardContent() {
-        return axios.get(API_URL + 'user', { headers: authHeader() });
+        return axios.get(API_URL + 'user', {headers: authHeader()});
     }
 
     getModeratorBoard() {
-        return axios.get(API_URL + 'mod', { headers: authHeader() });
+        return axios.get(API_URL + 'mod', {headers: authHeader()});
     }
 
     getAdminBoard() {
-        return axios.get(API_URL + 'admin', { headers: authHeader() });
+        return axios.get(API_URL + 'admin', {headers: authHeader()});
     }
 }
