+++ src/main/java/com/app/medicalwebapp/controllers/requestbody/messenger/MessagesRequest.java
@@ -1,6 +1,6 @@
 package com.app.medicalwebapp.controllers.requestbody.messenger;
 
-import com.app.medicalwebapp.model.messengerModels.ChatMessage;
+import com.app.medicalwebapp.model.messenger_models.ChatMessage;
 import lombok.Getter;
 
 import java.util.List;
