--- src/main/java/com/app/medicalwebapp/model/messengerModels/StatusMessage.java
@@ -1,6 +1,0 @@
-package com.app.medicalwebapp.model.messengerModels;
-
-public enum StatusMessage {
-    READ,
-    UNREAD
-}
