+++ src/main/java/com/app/medicalwebapp/model/messenger_models/StatusMessage.java
@@ -1,0 +1,6 @@
+package com.app.medicalwebapp.model.messenger_models;
+
+public enum StatusMessage {
+    READ,
+    UNREAD
+}
