+++ src/main/java/com/app/medicalwebapp/repositories/messenger_repositories/ChatMessageRepository.java
@@ -1,7 +1,7 @@
-package com.app.medicalwebapp.repositories.messengerRepositories;
+package com.app.medicalwebapp.repositories.messenger_repositories;
 
-import com.app.medicalwebapp.model.messengerModels.ChatMessage;
-import com.app.medicalwebapp.model.messengerModels.StatusMessage;
+import com.app.medicalwebapp.model.messenger_models.ChatMessage;
+import com.app.medicalwebapp.model.messenger_models.StatusMessage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
